Replaced json with custom protocol

+++ src/app/curves/simplify.ts
@@ -1,12 +1,12 @@
 export class Simplify {
 
-  private static GetSqDist(p1: Uint32Array, p2: Uint32Array): number {
+  private static GetSqDist(p1: Int16Array, p2: Int16Array): number {
     const dx = p1[0] - p2[0];
     const dy = p1[1] - p2[1];
     return dx * dx + dy * dy;
   }
 
-  private static GetSqSegDist(p: Uint32Array, p1: Uint32Array, p2: Uint32Array): number {
+  private static GetSqSegDist(p: Int16Array, p1: Int16Array, p2: Int16Array): number {
 
     let x = p1[0];
     let y = p1[1];
@@ -33,7 +33,7 @@ export class Simplify {
     return dx * dx + dy * dy;
   }
 
-  private static SimplifyRadialDist(points: Uint32Array[], sqTolerance: number) {
+  private static SimplifyRadialDist(points: Int16Array[], sqTolerance: number) {
 
     let prevPoint = points[0];
     const newPoints = [prevPoint];
@@ -53,7 +53,7 @@ export class Simplify {
     return newPoints;
   }
 
-  private static SimplifyDPStep(points: Uint32Array[], first: number, last: number, sqTolerance: number, simplified: Uint32Array[]): void {
+  private static SimplifyDPStep(points: Int16Array[], first: number, last: number, sqTolerance: number, simplified: Int16Array[]): void {
     let maxSqDist = sqTolerance;
     let index;
 
@@ -73,7 +73,7 @@ export class Simplify {
     }
   }
 
-  public static SimplifyDouglasPeucker(points: Uint32Array[], sqTolerance: number): Uint32Array[] {
+  public static SimplifyDouglasPeucker(points: Int16Array[], sqTolerance: number): Int16Array[] {
     const last = points.length - 1;
 
     const simplified = [points[0]];
@@ -83,7 +83,7 @@ export class Simplify {
     return simplified;
   }
 
-  public static Simplify(points: Uint32Array[], tolerance: number, highestQuality = false) {
+  public static Simplify(points: Int16Array[], tolerance: number, highestQuality = false) {
 
     if (points.length <= 2) { return points; }
